refactor(ava): refactor type
