--- src/common/types.ts
@@ -1,15 +1,0 @@
-/**
- * One row(record) of data in JSON.
- */
-export type Datum = Record<string, any>;
-
-/**
- * Rows(records) of data.
- */
-export type Data = Datum[];
-
-/**
- * Specification: declarative schema to describe a visualization.
- */
-export type Specification = any;
-export type ChartSpec = any;
