+++ src/types/render.ts
@@ -1,5 +1,10 @@
-import { Spec } from '../bind';
 import { CHART_PURPOSE } from '../constants/advisor';
+
+// We should export Spec and Renderer type here to avoid circular dependency between ava and gpt-vis.
+// TODO: import type { Spec } from '@antv/gpt-vis';
+export type Spec = any;
+
+export type Renderer = (container: string, spec: Spec) => void;
 
 export type TrendData = Array<{
   time: string;
